Moved utils into a unified folder
