+++ src/app/(home)/articles-section.tsx
@@ -1,6 +1,6 @@
 import { GoPencil } from "react-icons/go";
 import { HighlightCardRow } from "~/app/components/highlight-card-row";
-import { getAllPostsMeta } from "~/server/utils/mdx/mdx";
+import { getAllPostsMeta } from "~/utils/server/mdx/mdx";
 
 export default async function Articles() {
   const articles = await getAllPostsMeta("articles");
