+++ src/app/(mdx)/articles/page.tsx
@@ -1,6 +1,6 @@
 import AllPosts from "~/app/(mdx)/all-posts";
-import { projectToCardMatter } from "~/server/utils/mdx/converters";
-import { getAllPostsMeta } from "~/server/utils/mdx/mdx";
+import { projectToCardMatter } from "~/utils/server/mdx/converters";
+import { getAllPostsMeta } from "~/utils/server/mdx/mdx";
 
 export default async function Articles() {
   const allArticles = (await getAllPostsMeta("articles")).map(
