+++ src/app/(mdx)/mdx-layout.tsx
@@ -2,8 +2,8 @@ import { type PropsWithChildren } from "react";
 import MdxHeader from "~/app/(mdx)/mdx-header";
 import { MainContainer } from "~/app/components/mdx/main-container";
 import { PostsPanel } from "~/app/components/navigation/posts-panel";
-import { getAllPostsMeta } from "~/server/utils/mdx/mdx";
-import { type SourceRoute } from "~/server/utils/mdx/types";
+import { getAllPostsMeta } from "~/utils/server/mdx/mdx";
+import { type SourceRoute } from "~/utils/server/mdx/types";
 
 export default async function MdxLayout({
   children,
