+++ src/app/components/highlight-card.tsx
@@ -4,7 +4,7 @@ import { type ReactNode } from "react";
 import { Card } from "~/app/components/card";
 import { BadgeList } from "~/app/components/mdx/badge-list";
 import { Badge } from "~/app/components/ui/badge";
-import { cn } from "~/app/utils/styles";
+import { cn } from "~/utils/client/styles";
 
 export interface HighlightCardProps {
   title: string;
