+++ src/app/components/ui/badge.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "~/app/utils/styles";
+import { cn } from "~/utils/client/styles";
 
 const badgeVariants = cva(
   "inline-flex whitespace-nowrap items-center rounded-lg border border-neutral-200 px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-neutral-950 focus:ring-offset-2 dark:border-neutral-800 dark:focus:ring-neutral-300",
