+++ src/app/components/ui/popover.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import * as PopoverPrimitive from "@radix-ui/react-popover";
-import { cn } from "~/app/utils/styles";
+import { cn } from "~/utils/client/styles";
 
 const Popover = PopoverPrimitive.Root;
