+++ src/app/components/ui/sheet.tsx
@@ -5,7 +5,7 @@ import * as SheetPrimitive from "@radix-ui/react-dialog";
 import { cva, type VariantProps } from "class-variance-authority";
 import { X } from "lucide-react";
 
-import { cn } from "~/app/utils/styles";
+import { cn } from "~/utils/client/styles";
 
 const Sheet = SheetPrimitive.Root;
