+++ src/app/components/ui/toast.tsx
@@ -3,7 +3,7 @@ import * as ToastPrimitives from "@radix-ui/react-toast";
 import { cva, type VariantProps } from "class-variance-authority";
 import { X } from "lucide-react";
 
-import { cn } from "~/app/utils/styles";
+import { cn } from "~/utils/client/styles";
 
 const ToastProvider = ToastPrimitives.Provider;
